Backend/requirements.txt
beautifulsoup4
CacheControl
chardet
colorama
contextlib2
distlib
distro
docopt
html5lib
idna
lxml
mechanize
msgpack
npyscreen
numpy
ordered-set
packaging
paho-mqtt
pandas
pep517
phantomjs
Pillow
ply
progress
pycparser
pyparsing
pytz
PyYAML
selenium
soupsieve
toml
udiskie
webencodings
urllib3
requests
idna

